sdm_health_exporter.py - v0.0.1
This script serves as an example exporter that can monitor the
health of resources ("Infrastructure") and nodes ("Gateways/Relays").

The script uses the following workflow:

- Make an API call to strongDM's API to retrive information about resources and
nodes. The frequency of the API call is configurable by updating the "update_interval"
variable in "main()"

- Collect data about any resource or node that is tagged with <alert_tag> in strongDM.
This tag is configurable by updating the "alert_tag" variable in "main()"

- Export metrics to a prometheus endpoint as a "Gauge" (0 for healthy, 1 for unhealthy)